Some more refactor
