+++ twindle-cli/twitter/scraping/index.js
@@ -1,1 +1,1 @@
-// for the future once puppeteer is ready we can bring it into this folder
+// once puppeteer is ready we can bring it into this folder
