+++ twindle-cli/twitter/transformations/helpers.js
@@ -6,6 +6,8 @@
  */
 
 const twemoji = require("twemoji");
+const { matchAll } = require("../utils/string");
+const { format } = require("../utils/date");
 
 /**
  * Render the images, videos and GIFs
@@ -39,7 +41,9 @@ function renderMedia(tweetObj) {
 
   for (let mediaKey of mediaKeys) {
     // Search for it in the expandedMedia
-    const mediaInfo = expandedMediaIncludes.find(({ media_key }) => media_key === mediaKey);
+    const mediaInfo = expandedMediaIncludes.find(
+      ({ media_key }) => media_key === mediaKey
+    );
 
     const { width, height, url, preview_image_url } = mediaInfo;
 
@@ -50,7 +54,8 @@ function renderMedia(tweetObj) {
     const urls = tweetObj.entities.urls;
 
     const urlObjOfImage = urls.find(
-      ({ expanded_url }) => expanded_url.includes("/photo/") || expanded_url.includes("/video/")
+      ({ expanded_url }) =>
+        expanded_url.includes("/photo/") || expanded_url.includes("/video/")
     );
 
     // Add to our list
@@ -110,7 +115,9 @@ function renderOutsiderLinks(tweetObj) {
 
     if (isACard && !linkWithImageChosen && !hasCustomMedia) {
       // Check if it is at the end or not
-      if (tweetText.trim().substring(urlObj.start, urlObj.end + 1) === urlObj.url) {
+      if (
+        tweetText.trim().substring(urlObj.start, urlObj.end + 1) === urlObj.url
+      ) {
         // Remove from the markup
         tweetObj.text = tweetText.replace(urlObj.url, "");
       }
@@ -135,7 +142,8 @@ function renderOutsiderLinks(tweetObj) {
     );
   }
 
-  if (Object.entries(linkWithImage).length) tweetObj.linkWithImage = linkWithImage;
+  if (Object.entries(linkWithImage).length)
+    tweetObj.linkWithImage = linkWithImage;
 
   return tweetObj;
 }
@@ -172,7 +180,10 @@ function renderMentionsHashtags({ text = "", mentions = [], hashtags = [] }) {
       const { tag } = hashtag;
 
       // Replace
-      text = text.replace(`#${tag}`, `<a href="https://twitter.com/hashtag/${tag}">#${tag}</a>`);
+      text = text.replace(
+        `#${tag}`,
+        `<a href="https://twitter.com/hashtag/${tag}">#${tag}</a>`
+      );
     }
   }
 
@@ -205,7 +216,6 @@ function fixUserDescription(tweets) {
 
   const descriptionURLs = entitiesDescription && entitiesDescription.urls;
 
-
   if (!descriptionURLs) return tweets;
 
   for (let descriptionURLObj of descriptionURLs) {
@@ -237,4 +247,74 @@ function renderRichTweets(tweetObj) {
   return tweetObj;
 }
 
-module.exports = { renderRichTweets, fixUserDescription };
+/**
+ * Get tweet ID from URL `https://twitter.com/[USER]/status/[ID]` | Very flexible,
+ * Will obtain IDs from `status/[ID]/photo/1` or `[ID]?s=20`
+ * @param {string} tweet_url
+ * @returns {string}
+ */
+const extractTweetId = (tweet_url) =>
+  [
+    // @ts-ignore
+    ...matchAll(
+      tweet_url,
+      /https?:\/\/twitter.com\/[a-zA-Z_]{1,20}\/status\/([0-9]*)/g
+    ),
+  ][0][1];
+
+const extractScreenName = (tweet_url) =>
+  tweet_url
+    .substring(0, tweet_url.lastIndexOf("/status"))
+    .substring(tweet_url.lastIndexOf("/") + 1);
+
+const getTweetArray = (responseJSON) => {
+  return (responseJSON.data || []).map((data) => ({
+    ...data,
+    includes: responseJSON.includes,
+  }));
+};
+
+const getUserObject = (responseJSON) => responseJSON.includes.users[0];
+
+const getTweetObject = (responseJSON) => ({
+  ...responseJSON.data[0],
+  includes: responseJSON.includes,
+});
+
+const createCustomTweet = (tweet_object, user_object) => {
+  // if (!tweet_object) return {};
+  // console.log({ tweet_object });
+  return {
+    id: tweet_object.id,
+    createdAt: format(
+      new Date(tweet_object.created_at),
+      "MMM d, yyyy  h:mm aaaa"
+    ),
+    tweet: twemoji.parse(fixLineBreaks(tweet_object.text), {
+      folder: "svg",
+      ext: ".svg",
+    }),
+    customMedia: tweet_object.customMedia,
+    linkWithImage: tweet_object.linkWithImage,
+  };
+};
+
+/**
+ * Replaces `\n` with `<br />`
+ * @param {string} tweet
+ */
+function fixLineBreaks(tweet) {
+  return tweet.replace(/\n/g, "<br />");
+}
+
+module.exports = {
+  renderRichTweets,
+  fixUserDescription,
+  extractTweetId,
+  extractScreenName,
+  getTweetArray,
+  getUserObject,
+  createCustomTweet,
+  fixLineBreaks,
+  getTweetObject,
+};
