+++ twindle-cli/twitter/utils/date.js
@@ -1,0 +1,5 @@
+const { format } = require("date-fns");
+
+module.exports = {
+  format,
+};
