+++ twindle-cli/twitter/utils/string.js
@@ -1,0 +1,5 @@
+const matchAll = require("string.prototype.matchall");
+
+module.exports = {
+  matchAll,
+};
